requirements.txt
blinker==1.9.0
click==8.3.2
et-xmlfile==2.0.0
flask==3.1.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
greenlet==3.4.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
openpyxl==3.1.5
pyyaml==6.0.3
sqlalchemy==2.0.49
typing-extensions==4.15.0
werkzeug==3.1.8
wtforms==3.2.1
